Update README.md
Briefly summarize the DriverPass project. Who was the client? What type of system did they want you to design?
The driver Pass wants to provide their users. To have access to practice tests and to schedule or change their appointment. The driver could leave notes on the user web page. The owner wanted to be able to keep track of everything that gets changed on the website and to be able to make reports with the data.
What did you do particularly well?
I feel that my diagrams were easy to understand and provided a visualization of the critical concepts of the application.
If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?
I would have included more limitations and assumptions.
How did you interpret the user’s needs and implement them into your system design? Why is it so important to consider the user’s needs when designing?
I felt that understanding the user’s needs and requirements from the interview. I used luidcharts to make diagrams that showed the user’s needs and to adapt the needs what would work well for the website. The front-line users use the website the most so their needs are the most important.
How do you approach designing software? What techniques or strategies would you use in the future to analyze and create a system?
I feel like using luidcarts. Making the case user, UML, and the Gantt chart made it very helpful and more accessible to visualize and give the others a better understanding of what is going on.